moved test client fixture to fixture_artifact file
