+++ tests/fixtures/fixture_artifact.py
@@ -1,8 +1,11 @@
 import pytest 
+from app import app
 
-@pytest.fixture(scope="module")
-def sample_fixture():
-    city = ["NYC", "Chicago", "London", "Mumbai"]
-    yield city
+@pytest.fixture
+def client():
+    """Creating the test client"""
+    app.config['TESTING'] = True
+    with app.test_client() as client:
+        yield client
