+++ tests/test_functions/test_view_ledger.py
@@ -5,15 +5,6 @@ from unittest.mock import patch, MagicMock
 import os
 import sys
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "../.."))
-from app import app
-
-
-@pytest.fixture
-def client():
-    """Creating the test client"""
-    app.config['TESTING'] = True
-    with app.test_client() as client:
-        yield client
 
 
 @patch('app.get_db_connection')
